Restore confirmation dialog for track deletion

+++ src/pages/SingleTrack.js
@@ -34,7 +34,7 @@ class SingleTrack extends Component {
     }
 
     deleteTrack() {
-        // if(!confirm('Are you sure you want to delete this track?')) { return; }
+        if(!confirm('Are you sure you want to delete this track?')) { return; }
         this.props.trackStore.deleteTrack(parseInt(this.props.params.trackId, 10))
         .then(() => { this.props.router.push('/tracks'); });
     }
